4.21*

+++ controllers/bloglists.js
@@ -22,7 +22,7 @@ blogsRouter.post('/', async (request, response) => {
      const body = request.body
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
     if(!request.token || !decodedToken.id){
-        return response.status(401).json({error: 'token missing r invalid'})
+        return response.status(401).json({error: 'token missing or invalid'})
     }
     const user = await User.findById(decodedToken.id)
     const blog = new Blog({
@@ -40,8 +40,12 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if(!request.token || !decodedToken){
+        return response.status(401).json({error: "token is missing or invalid"})
+    }
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
 })
 
 blogsRouter.put('/:id', async (request, response) => {
